requirements.txt
appnope==0.1.0
beautifulsoup4==4.5.1
bleach==1.5.0
bs4==0.0.1
chainer==1.19.0
cycler==0.10.0
decorator==4.0.10
entrypoints==0.2.2
falcon==1.1.0
filelock==2.0.7
future==0.16.0
gunicorn==19.6.0
h5py==2.6.0
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
line-bot-sdk==1.1.0
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
nbconvert==5.0.0
nbformat==4.2.0
nose==1.3.7
notebook==4.3.1
numpy==1.11.3
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
ptyprocess==0.5.1
py-nifty-cloud==0.9.4
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
widgetsnbextension==1.2.6
